+++ greenguard/loaders/__init__.py
@@ -1,0 +1,5 @@
+from greenguard.loaders.csv import CSVLoader
+
+__all__ = (
+    'CSVLoader',
+)