+++ HISTORY.md
@@ -1,5 +1,14 @@
 # History
 
+## 0.2.0 - 2020-02-14
+
+First stable release:
+
+* efficient data loading and preprocessing
+* initial collection of dfs and lstm based pipelines
+* optimized pipeline tuning
+* documentation and tutorials
+
 ## 0.1.0
 
 * First release on PyPI